Cast slightly abused enum to int, so that GCC 4.5 won't warn about switch cases that are not part of the enum's definition.

+++ src/backend/nodes/read.c
@@ -289,7 +289,7 @@ nodeRead(char *token, int tok_len)
 
 	type = nodeTokenType(token, tok_len);
 
-	switch (type)
+	switch ((int) type)
 	{
 		case LEFT_BRACE:
 			result = parseNodeString();
